Remove my-revert-buffer and associated key binding
I'm trying to wean myself off of certain non standard keybindings. The default
revert-buffer is bound to s-u, so I want to start using that. More recently,
revert-buffer-quick was added to Emacs, which is bound to C-x x g. Except for
preserving modes, the latter looks like a good replacement for my-revert-buffer.